Monorepo: address circular dependencies surfaced by SWC TDZ.
