+++ packages/js/components/src/dynamic-form/dynamic-form.tsx
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { Form } from '../index';
+import { Form } from '../form';
 import {
 	TextField,
 	PasswordField,
