+++ packages/js/navigation/src/index.js
@@ -2,22 +2,21 @@
  * External dependencies
  */
 import { useState, useEffect, useLayoutEffect } from '@wordpress/element';
-import { addQueryArgs } from '@wordpress/url';
-import { parse } from 'qs';
 import { pick } from 'lodash';
 import { applyFilters } from '@wordpress/hooks';
-import { getAdminLink } from '@woocommerce/settings';
 
 /**
  * Internal dependencies
  */
 import { getHistory } from './history';
+import { getPath, getQuery, getNewPath, parseAdminUrl } from './url';
 
 // Expose history so all uses get the same history object.
 export { getHistory };
 
 // Export all filter utilities
 export * from './filters';
+export { getPath, getQuery, getNewPath, parseAdminUrl } from './url';
 
 // Export all hooks
 export { useConfirmUnsavedChanges } from './hooks/use-confirm-unsaved-changes';
@@ -25,48 +24,6 @@ export { useConfirmUnsavedChanges } from './hooks/use-confirm-unsaved-changes';
 const TIME_EXCLUDED_SCREENS_FILTER = 'woocommerce_admin_time_excluded_screens';
 const NAVIGATION_UPDATE_EXCLUDED_SCREENS_FILTER =
 	'woocommerce_admin_nav_update_excluded_screens';
-
-/**
- * Get the current path from history.
- *
- * @return {string}  Current path.
- */
-export const getPath = () => getHistory().location.pathname;
-
-/**
- * Get the current query string, parsed into an object, from history.
- *
- * @return {Object}  Current query object, defaults to empty object.
- */
-export function getQuery() {
-	const search = getHistory().location.search;
-	if ( search.length ) {
-		return parse( search.substring( 1 ) ) || {};
-	}
-	return {};
-}
-
-/**
- * Return a URL with set query parameters.
- *
- * @param {Object} query        object of params to be updated.
- * @param {string} path         Relative path (defaults to current path).
- * @param {Object} currentQuery object of current query params (defaults to current querystring).
- * @param {string} page         Page key (defaults to "wc-admin")
- * @return {string}  Updated URL merging query params into existing params.
- */
-export function getNewPath(
-	query,
-	path = getPath(),
-	currentQuery = getQuery(),
-	page = 'wc-admin'
-) {
-	const args = { page, ...currentQuery, ...query };
-	if ( path !== '/' ) {
-		args.path = path;
-	}
-	return addQueryArgs( 'admin.php', args );
-}
 
 /**
  * Gets query parameters that should persist between screens or updates
@@ -311,22 +268,6 @@ export function onQueryChange( param, path = getPath(), query = getQuery() ) {
  */
 export const isWCAdmin = ( url = window.location.href ) => {
 	return /admin.php\?page=wc-admin/.test( url );
-};
-
-/**
- * Returns a parsed object for an absolute or relative admin URL.
- *
- * @param {*} url - the url to test.
- * @return {URL} - the URL object of the given url.
- */
-export const parseAdminUrl = ( url ) => {
-	if ( url.startsWith( 'http' ) ) {
-		return new URL( url );
-	}
-
-	return /^\/?[a-z0-9]+.php/i.test( url )
-		? new URL( `${ window.wcSettings.adminUrl }${ url }` )
-		: new URL( getAdminLink( getNewPath( {}, url, {} ) ) );
 };
 
 /**
