+++ plugins/woocommerce/client/admin/client/layout/test/controller.test.js
@@ -8,6 +8,14 @@ import * as navigation from '@woocommerce/navigation';
  */
 import { updateLinkHref } from '../controller';
 
+jest.mock( '@woocommerce/navigation', () => {
+	const actual = jest.requireActual( '@woocommerce/navigation' );
+	return {
+		...actual,
+		getHistory: jest.fn( actual.getHistory ),
+	};
+} );
+
 describe( 'updateLinkHref', () => {
 	const timeExcludedScreens = [ 'stock', 'settings', 'customers' ];
 
@@ -27,6 +35,7 @@ describe( 'updateLinkHref', () => {
 
 	beforeEach( () => {
 		jest.restoreAllMocks();
+		navigation.getHistory.mockClear();
 	} );
 
 	it( 'should update report urls', () => {
@@ -88,7 +97,6 @@ describe( 'updateLinkHref', () => {
 
 	it( 'should not prevent default when Command key is pressed', () => {
 		const item = { href: REPORT_URL };
-		const spyGetHistory = jest.spyOn( navigation, 'getHistory' );
 		const event = {
 			ctrlKey: false,
 			metaKey: true,
@@ -98,13 +106,12 @@ describe( 'updateLinkHref', () => {
 		updateLinkHref( item, nextQuery, timeExcludedScreens );
 
 		item.onclick( event );
-		expect( spyGetHistory ).not.toHaveBeenCalled();
+		expect( navigation.getHistory ).not.toHaveBeenCalled();
 		expect( event.preventDefault ).not.toHaveBeenCalled();
 	} );
 
 	it( 'should not prevent default when Control key is pressed', () => {
 		const item = { href: REPORT_URL };
-		const spyGetHistory = jest.spyOn( navigation, 'getHistory' );
 		const event = {
 			ctrlKey: true,
 			metaKey: false,
@@ -114,13 +121,12 @@ describe( 'updateLinkHref', () => {
 		updateLinkHref( item, nextQuery, timeExcludedScreens );
 
 		item.onclick( event );
-		expect( spyGetHistory ).not.toHaveBeenCalled();
+		expect( navigation.getHistory ).not.toHaveBeenCalled();
 		expect( event.preventDefault ).not.toHaveBeenCalled();
 	} );
 
 	it( 'should prevent default on normal clicks', () => {
 		const item = { href: REPORT_URL };
-		const spyGetHistory = jest.spyOn( navigation, 'getHistory' );
 		const event = {
 			ctrlKey: false,
 			metaKey: false,
@@ -130,7 +136,7 @@ describe( 'updateLinkHref', () => {
 		updateLinkHref( item, nextQuery, timeExcludedScreens );
 
 		item.onclick( event );
-		expect( spyGetHistory ).toHaveBeenCalledTimes( 1 );
+		expect( navigation.getHistory ).toHaveBeenCalledTimes( 1 );
 		expect( event.preventDefault ).toHaveBeenCalledTimes( 1 );
 	} );
 } );
